Hard coded navigation for non-JS users
The hard coded navigation works in 'en' and translations, but if JS is
disable it is shown in English in sub folders, links work though as a
sub folder ('es', 'ja'...).

If JS is enable it automatically takes the nav for the selected
language, with caption well translated.

Now non-JS users can now see the navigation menu, BUT not the languages
menu as it need to be changed for each lang addition, on each page, for
each translation. It would mean a unmaintainable way to keep the langs
menu. So keeping it JS-only won't hurt.